requirements.txt
fastapi
uvicorn
sqlmodel
python-multipart
pydantic
python-jose[cryptography]
bcrypt
langchain
langchain-community
langchain-groq
langgraph
python-dotenv
pandas
pypdf
python-pptx
python-docx
chromadb
camelot-py
opencv-python
tabulate
langchain-huggingface
langchain-pinecone
pinecone-client
httpx
